Extract some code into new file to avoid circular imports
